% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{benchmark_backends}
\alias{benchmark_backends}
\title{Benchmark backends}
\usage{
benchmark_backends(tree, Y, n_reps = 10)
}
\arguments{
\item{tree}{A phylo object}

\item{Y}{Data frame with traits}

\item{n_reps}{Number of repetitions for timing}
}
\value{
Data frame with timing results
}
\description{
Compare performance of different backends on the same data.
}
