% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_block_constraint}
\alias{build_block_constraint}
\title{Build block-diagonal constraint}
\usage{
build_block_constraint(
  p,
  blocks,
  is_correlated,
  fixed,
  shared_var,
  matrix_name
)
}
\description{
Build block-diagonal constraint
}
\keyword{internal}
