% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_diagonal_constraint}
\alias{build_diagonal_constraint}
\title{Build diagonal-only constraint}
\usage{
build_diagonal_constraint(p, fixed, shared_var, matrix_name)
}
\description{
Build diagonal-only constraint
}
\keyword{internal}
