% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_full_constraint}
\alias{build_full_constraint}
\title{Build full (correlated) constraint}
\usage{
build_full_constraint(p, fixed, shared_var, matrix_name)
}
\description{
Build full (correlated) constraint
}
\keyword{internal}
