% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_matrix_constraint}
\alias{build_matrix_constraint}
\title{Build constraint specification for one matrix}
\usage{
build_matrix_constraint(
  p,
  is_correlated = TRUE,
  fixed = NULL,
  shared_var = NULL,
  blocks = NULL,
  matrix_name = "R"
)
}
\arguments{
\item{p}{Dimension}

\item{is_correlated}{If FALSE, diagonal-only (no off-diagonal)}

\item{fixed}{List of fixed entry specs}

\item{shared_var}{List of shared variance groups}

\item{blocks}{List of block-diagonal groups}

\item{matrix_name}{"R" or "S" (for error messages)}
}
\value{
List describing the constraint
}
\description{
Build constraint specification for one matrix
}
\keyword{internal}
