% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_grad_block}
\alias{constrained_grad_block}
\title{Compute constrained gradient for block-diagonal matrix}
\usage{
constrained_grad_block(grad_mat, mat, spec)
}
\description{
Compute constrained gradient for block-diagonal matrix
}
\keyword{internal}
