% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_grad_one_matrix}
\alias{constrained_grad_one_matrix}
\title{Compute constrained gradient for one matrix}
\usage{
constrained_grad_one_matrix(grad_mat, mat, spec)
}
\description{
Compute constrained gradient for one matrix
}
\keyword{internal}
