% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_gradient_to_theta}
\alias{constrained_gradient_to_theta}
\title{Convert constrained gradient matrices to theta-space gradient}
\usage{
constrained_gradient_to_theta(grad_R, grad_S, R, S, constraints)
}
\arguments{
\item{grad_R}{Gradient w.r.t. R matrix (p x p)}

\item{grad_S}{Gradient w.r.t. S matrix (p x p)}

\item{R}{Current R matrix}

\item{S}{Current S matrix}

\item{constraints}{ConstraintSpec object}
}
\value{
Numeric vector of gradient in constrained theta space
}
\description{
Convert constrained gradient matrices to theta-space gradient
}
\keyword{internal}
