% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_matrices_to_theta}
\alias{constrained_matrices_to_theta}
\title{Convert R and S matrices to constrained theta vector}
\usage{
constrained_matrices_to_theta(R, S, constraints)
}
\arguments{
\item{R}{Phylogenetic rate matrix}

\item{S}{Residual covariance matrix}

\item{constraints}{ConstraintSpec object}
}
\value{
Numeric vector of free parameters
}
\description{
Convenience function: given R and S matrices, produce the combined
constrained theta vector.
}
\keyword{internal}
