% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_matrix_to_theta}
\alias{constrained_matrix_to_theta}
\title{Convert covariance matrix to constrained theta}
\usage{
constrained_matrix_to_theta(mat, spec)
}
\arguments{
\item{mat}{p x p covariance matrix}

\item{spec}{Constraint spec for one matrix}
}
\value{
Numeric vector of free parameters
}
\description{
Convert covariance matrix to constrained theta
}
\keyword{internal}
