% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_matrix_to_theta_block}
\alias{constrained_matrix_to_theta_block}
\title{Convert block-diagonal matrix to constrained theta}
\usage{
constrained_matrix_to_theta_block(mat, spec)
}
\description{
Convert block-diagonal matrix to constrained theta
}
\keyword{internal}
