% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constrained_theta_to_matrix_block}
\alias{constrained_theta_to_matrix_block}
\title{Convert constrained theta to block-diagonal matrix}
\usage{
constrained_theta_to_matrix_block(theta, spec)
}
\description{
Convert constrained theta to block-diagonal matrix
}
\keyword{internal}
