% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_from_root}
\alias{dist_from_root}
\title{Calculate distance from root to each node}
\usage{
dist_from_root(tree)
}
\arguments{
\item{tree}{A phylo object}
}
\value{
The tree with added 'root_dist' element
}
\description{
Calculate distance from root to each node
}
