% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_fit_eigen}
\alias{em_fit_eigen}
\title{EM Algorithm - Eigen}
\usage{
em_fit_eigen(
  R_init,
  S_init,
  mu_init,
  edge,
  edge_length,
  traits,
  species_idx,
  n_tips,
  n_nodes,
  max_iter = 100L,
  tol = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{R_init}{Initial phylogenetic rate matrix}

\item{S_init}{Initial residual covariance matrix}

\item{mu_init}{Initial mean vector}

\item{edge}{Edge matrix (n_edges x 2), 1-indexed}

\item{edge_length}{Edge lengths}

\item{traits}{Trait matrix (nind x p)}

\item{species_idx}{Species index for each observation (1-indexed)}

\item{n_tips}{Number of tips}

\item{n_nodes}{Total nodes}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
List with R, S, mu, logL, converged, iterations
}
\description{
Full EM algorithm (Eigen implementation).
}
