% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tree}
\alias{example_tree}
\title{Example phylogenetic tree}
\format{
A phylo object with 25 tips
}
\usage{
example_tree
}
\description{
A phylogenetic tree with 25 species for demonstrating phylopars functions.
}
\examples{
data(example_tree)
plot(example_tree)
}
\keyword{datasets}
