% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_starting_theta}
\alias{get_default_starting_theta}
\title{Generate default starting values for optimization}
\usage{
get_default_starting_theta(
  m,
  Sigma_diag = FALSE,
  B_diag = FALSE,
  Sigma_fixed = NULL,
  B_fixed = NULL,
  Sigma_start = NULL,
  B_start = NULL,
  model = "BM",
  evo_model_par_start = NULL,
  evo_model_par_fixed = NULL
)
}
\arguments{
\item{m}{Dimension of covariance matrices}

\item{Sigma_diag}{Logical, constrain Sigma to diagonal}

\item{B_diag}{Logical, constrain B to diagonal}

\item{Sigma_fixed}{Fixed Sigma matrix}

\item{B_fixed}{Fixed B matrix}

\item{Sigma_start}{Starting value for Sigma}

\item{B_start}{Starting value for B}

\item{model}{Evolutionary model}

\item{evo_model_par_start}{Starting value for evolutionary parameter}

\item{evo_model_par_fixed}{Fixed evolutionary parameter}
}
\value{
Numeric vector of starting parameter values
}
\description{
Generate default starting values for optimization
}
