% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Analytic Gradient for Phylogenetic Mixed Model}
\description{
Computes the analytic gradient of the log-likelihood with respect to the
phylogenetic rate matrix (R/Sigma) and residual covariance matrix (S/B)
using the Bastide 2021 algorithm.
}
\details{
The gradient formulas (from Bastide et al. 2021) are:

For the phylogenetic rate R (summed over internal edges k):
  dlogL/dR = sum_k -0.5 * t_k * Q_k %*% (Q_k^{-1} - (M_k - qn_k)(M_k - qn_k)' - V_k) %*% Q_k

For the residual covariance S (summed over observations k):
  dlogL/dS = sum_k -0.5 * Q_k %*% (Q_k^{-1} - (M_k - qn_k)(M_k - qn_k)' - V_k) %*% Q_k

Where:
- Q_k is the backward precision matrix at node k
- M_k is the conditional mean (ASR estimate)
- qn_k is the backward conditional mean
- V_k is the conditional variance
- t_k is the edge length
}
