% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_bastide_arma}
\alias{loglik_bastide_arma}
\title{Bastide Traversal (Forward Only) - Armadillo}
\usage{
loglik_bastide_arma(
  R,
  S,
  edge,
  edge_length,
  traits,
  species_idx,
  n_tips,
  n_nodes
)
}
\arguments{
\item{R}{Phylogenetic rate matrix (p x p)}

\item{S}{Residual covariance matrix (p x p)}

\item{edge}{Edge matrix from phylo object (n_edges x 2), 1-indexed}

\item{edge_length}{Edge lengths vector}

\item{traits}{Trait matrix (nind x p), NA for missing}

\item{species_idx}{Species index for each observation (1-indexed tip node)}

\item{n_tips}{Number of tip species}

\item{n_nodes}{Total number of nodes (tips + internal)}
}
\value{
ML log-likelihood value
}
\description{
Computes ML log-likelihood using Bastide 2021 algorithm.
This is the forward pass only (tips to root).
}
