% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends_advanced.R
\name{loglik_bastide_cpp}
\alias{loglik_bastide_cpp}
\title{Compute Log-Likelihood Using C++ Bastide Traversal}
\usage{
loglik_bastide_cpp(R, S, tree, trait_data, backend = "armadillo")
}
\arguments{
\item{R}{Phylogenetic rate matrix (p x p)}

\item{S}{Residual covariance matrix (p x p)}

\item{tree}{A phylo object (should be postorder)}

\item{trait_data}{Data frame with species in first column, traits in rest}

\item{backend}{Which backend to use: "armadillo" or "eigen"}
}
\value{
ML log-likelihood value
}
\description{
Fast C++ implementation of the Bastide 2021 ML likelihood computation.
}
