% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loo_trees}
\alias{loo_trees}
\title{50 LOO-CV Trees}
\format{
A list of phylo objects
}
\usage{
loo_trees
}
\description{
A named list of phylogenetic trees, each corresponding to a tree where a single species was dropped from `tree`.
The order of dropped species corresponds to the species vector `tree$tip.label`.
}
\examples{
data(loo_trees)
plot(loo_trees$t1)
}
\keyword{datasets}
