% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loo_Ys}
\alias{loo_Ys}
\title{50 LOO-CV Y Datasets}
\format{
A list of data frames
}
\usage{
loo_Ys
}
\description{
A named list of data frames, each corresponding to a subset of Y where all rows for a single species were removed.
The order of dropped species corresponds to the species vector `tree$tip.label`.
}
\examples{
data(loo_Ys)
head(loo_Ys$t1)
}
\keyword{datasets}
