% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_constraints}
\alias{make_constraints}
\title{Create a Constraint Specification}
\usage{
make_constraints(
  p,
  phylo_correlated = TRUE,
  pheno_correlated = TRUE,
  R_fixed = NULL,
  S_fixed = NULL,
  R_shared_var = NULL,
  S_shared_var = NULL,
  R_blocks = NULL,
  S_blocks = NULL
)
}
\arguments{
\item{p}{Integer, dimension of the covariance matrices (number of traits)}

\item{phylo_correlated}{Logical, if FALSE constrains R to be diagonal (default TRUE)}

\item{pheno_correlated}{Logical, if FALSE constrains S to be diagonal (default TRUE)}

\item{R_fixed}{List of lists, each with elements i, j, value. Fixes R[i,i] to
a known value by constraining the corresponding Cholesky elements. Only diagonal
entries (i == j) are supported. When R[i,i] is fixed to value v, L[i,i] = sqrt(v)
and L[i,j] = 0 for j < i, which also forces R[i,j] = 0 for j < i.}

\item{S_fixed}{List of lists, each with elements i, j, value. Same as R_fixed but
for the S matrix.}

\item{R_shared_var}{List of integer vectors. Each vector specifies a group of
diagonal indices that share the same L diagonal element (exp(theta)).
E.g., list(c(1,2,3)) means L[1,1] = L[2,2] = L[3,3] = exp(theta_shared).
Note: this shares the Cholesky diagonal, so R[i,i] values may differ slightly
due to off-diagonal L elements (R[i,i] = sum(L[i,:]^2)).}

\item{S_shared_var}{List of integer vectors, same as R_shared_var but for S.}

\item{R_blocks}{List of integer vectors specifying block-diagonal structure.
Each vector lists the trait indices in one block. Traits in different blocks
have zero covariance. E.g., list(1:3, 4:5) creates two independent blocks.}

\item{S_blocks}{List of integer vectors, same as R_blocks but for S.}
}
\value{
A list with class "ConstraintSpec" containing:
  \item{p}{Dimension of matrices}
  \item{R}{List with constraint details for R matrix}
  \item{S}{List with constraint details for S matrix}
  \item{n_free_R}{Number of free parameters for R}
  \item{n_free_S}{Number of free parameters for S}
  \item{n_free}{Total number of free matrix parameters}
}
\description{
Builds an internal representation of parameter constraints for the R
(phylogenetic rate) and S (within-species covariance) matrices.
}
