% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrices_to_theta}
\alias{matrices_to_theta}
\title{Convert R and S matrices to theta vector (Cholesky parameterization)}
\usage{
matrices_to_theta(R, S)
}
\description{
Convert R and S matrices to theta vector (Cholesky parameterization)
}
\keyword{internal}
