% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastide.R
\name{plogdet}
\alias{plogdet}
\title{Log Pseudo-Determinant}
\usage{
plogdet(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A matrix}

\item{tol}{Tolerance for determining positive eigenvalues}
}
\value{
Log pseudo-determinant
}
\description{
Computes the log of the product of positive singular values.
}
\keyword{internal}
