% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anc_recon.R
\name{print.anc_recon}
\alias{print.anc_recon}
\title{Print method for anc_recon objects}
\usage{
\method{print}{anc_recon}(x, ...)
}
\arguments{
\item{x}{An anc_recon object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for anc_recon objects
}
