% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopars.R
\name{print.phylopars}
\alias{print.phylopars}
\title{Print method for phylopars results}
\usage{
\method{print}{phylopars}(x, ...)
}
\arguments{
\item{x}{Result from phylopars()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for phylopars results
}
