% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{restricted_log_likelihood}
\alias{restricted_log_likelihood}
\title{Restricted log-likelihood for phylogenetic mixed model}
\usage{
restricted_log_likelihood(
  theta,
  tree,
  Y,
  ret_logL = TRUE,
  Sigma_diag = FALSE,
  B_diag = FALSE,
  Sigma_fixed = NULL,
  B_fixed = NULL,
  precompute_list = NULL,
  model = "BM",
  evo_model_par_fixed = NULL
)
}
\arguments{
\item{theta}{Parameter vector}

\item{tree}{A phylo object}

\item{Y}{Data frame with species in first column, traits in remaining columns}

\item{ret_logL}{If TRUE, return only the log-likelihood value}

\item{Sigma_diag}{Logical, if TRUE Sigma is diagonal}

\item{B_diag}{Logical, if TRUE B is diagonal}

\item{Sigma_fixed}{Fixed Sigma matrix}

\item{B_fixed}{Fixed B matrix}

\item{precompute_list}{List of precomputed values for efficiency}

\item{model}{Evolutionary model}

\item{evo_model_par_fixed}{Fixed evolutionary parameter}
}
\value{
Log-likelihood value or list with additional information
}
\description{
Calculates the REML log-likelihood using a postorder tree traversal algorithm.
This is O(n) in the number of species, handling missing data and within-species
variation.
}
