% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{restricted_loglik}
\alias{restricted_loglik}
\title{Backend-aware log-likelihood computation}
\usage{
restricted_loglik(
  theta,
  tree,
  Y,
  backend = c("R", "armadillo", "eigen"),
  Sigma_diag = FALSE,
  B_diag = FALSE,
  Sigma_fixed = NULL,
  B_fixed = NULL,
  precompute_list = NULL,
  model = "BM",
  evo_model_par_fixed = NULL
)
}
\arguments{
\item{theta}{Parameter vector}

\item{tree}{A phylo object}

\item{Y}{Data frame with species in first column, traits in remaining columns}

\item{backend}{Which backend to use: "R", "armadillo", or "eigen"}

\item{Sigma_diag}{Logical, if TRUE Sigma is diagonal}

\item{B_diag}{Logical, if TRUE B is diagonal}

\item{Sigma_fixed}{Fixed Sigma matrix}

\item{B_fixed}{Fixed B matrix}

\item{precompute_list}{List of precomputed values}

\item{model}{Evolutionary model}

\item{evo_model_par_fixed}{Fixed evolutionary parameter}
}
\value{
Log-likelihood value
}
\description{
Computes restricted log-likelihood using the specified backend.
}
