% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{restricted_loglik_arma}
\alias{restricted_loglik_arma}
\title{Restricted log-likelihood using RcppArmadillo}
\usage{
restricted_loglik_arma(
  Sigma,
  B,
  edge,
  edge_length,
  traits,
  species_idx,
  n_obs,
  Nnode
)
}
\arguments{
\item{Sigma}{Phylogenetic covariance matrix (m x m)}

\item{B}{Residual covariance matrix (m x m)}

\item{edge}{Edge matrix from phylo object (n_edges x 2), 1-indexed}

\item{edge_length}{Edge lengths vector}

\item{traits}{Trait matrix (nind x m), NA for missing}

\item{species_idx}{Species index for each observation (1-indexed tip node)}

\item{n_obs}{Total number of non-missing observations}

\item{Nnode}{Total number of nodes (tips + internal)}
}
\value{
Log-likelihood value, or -1e100 if singular
}
\description{
Fast C++ implementation of the postorder tree traversal algorithm.
}
