% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_matrix.R
\name{restricted_loglik_full}
\alias{restricted_loglik_full}
\title{Fit Model Using Full Matrix Approach}
\usage{
restricted_loglik_full(
  theta,
  tree,
  trait_data,
  m = NULL,
  Sigma_diag = FALSE,
  B_diag = FALSE,
  Sigma_fixed = NULL,
  B_fixed = NULL,
  model = "BM",
  evo_model_par_fixed = NULL
)
}
\arguments{
\item{theta}{Parameter vector}

\item{tree}{A phylo object}

\item{trait_data}{Data frame with species and traits}

\item{m}{Number of traits}

\item{Sigma_diag}{Constrain Sigma to diagonal}

\item{B_diag}{Constrain B to diagonal}

\item{Sigma_fixed}{Fixed value for Sigma}

\item{B_fixed}{Fixed value for B}

\item{model}{Evolutionary model}

\item{evo_model_par_fixed}{Fixed evolutionary parameter}
}
\value{
Log-likelihood value
}
\description{
Wrapper around loglik_full_matrix for use with optim().
Transforms theta to Sigma and B, then computes likelihood.
}
