% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{run_bfgs_step}
\alias{run_bfgs_step}
\title{Run BFGS Step with Analytic Gradients}
\usage{
run_bfgs_step(tree, trait_data, R, S, ctrl, verbose, method = "BFGS")
}
\description{
Run BFGS Step with Analytic Gradients
}
\keyword{internal}
