% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta2vals}
\alias{theta2vals}
\title{Convert parameter vector theta to covariance matrices}
\usage{
theta2vals(
  theta,
  m = NULL,
  Sigma_diag = FALSE,
  B_diag = FALSE,
  Sigma_fixed = NULL,
  B_fixed = NULL,
  model = "BM",
  evo_model_par_fixed = NULL
)
}
\arguments{
\item{theta}{Numeric vector of parameters}

\item{m}{Integer, dimension of the covariance matrices}

\item{Sigma_diag}{Logical, if TRUE Sigma is constrained to be diagonal}

\item{B_diag}{Logical, if TRUE B is constrained to be diagonal}

\item{Sigma_fixed}{Matrix or NULL, fixed value for Sigma (not optimized)}

\item{B_fixed}{Matrix or NULL, fixed value for B (not optimized)}

\item{model}{Character, evolutionary model ("BM" or "lambda")}

\item{evo_model_par_fixed}{Numeric or NULL, fixed evolutionary parameter}
}
\value{
List with components: Sigma, B, evo_model_par
}
\description{
Converts an unconstrained parameter vector to positive-definite covariance
matrices using Cholesky decomposition parameterization.
}
