% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transf_branch_lengths}
\alias{transf_branch_lengths}
\title{Transform branch lengths for evolutionary models}
\usage{
transf_branch_lengths(tree, model, evo_model_par)
}
\arguments{
\item{tree}{A phylo object with root_dist computed}

\item{model}{Evolutionary model ("BM" or "lambda")}

\item{evo_model_par}{Parameter for the model}
}
\value{
Tree with transformed branch lengths
}
\description{
Transform branch lengths for evolutionary models
}
