% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Example 50-species phylogenetic tree}
\format{
A phylo object with 50 tips
}
\usage{
tree
}
\description{
A phylogenetic tree with 50 species for demonstrating phylopars functions.
}
\examples{
data(tree)
plot(tree)
}
\keyword{datasets}
