% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_transforms.R
\name{validate_evo_par}
\alias{validate_evo_par}
\title{Validate Evolutionary Model Parameter}
\usage{
validate_evo_par(model, par, par_name = "parameter")
}
\arguments{
\item{model}{Character string: model name}

\item{par}{Numeric parameter value}

\item{par_name}{Character string: name of parameter (for error messages)}
}
\description{
Checks that a model parameter is within valid bounds.
}
\keyword{internal}
