% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gradient_to_theta}
\alias{gradient_to_theta}
\title{Convert gradient matrices to theta-space gradient}
\usage{
gradient_to_theta(grad_R, grad_S, R, S)
}
\description{
Uses chain rule: d/d(theta) = d/d(L) * d(L)/d(theta)
}
\keyword{internal}
