% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{run_bfgs_step}
\alias{run_bfgs_step}
\title{Run BFGS Step with Analytic Gradients}
\usage{
run_bfgs_step(
  tree,
  trait_data,
  R,
  S,
  ctrl,
  verbose,
  optim_method = "BFGS",
  likelihood = "ML",
  model = "BM",
  evo_model_par = NULL,
  optimize_evo = FALSE,
  constraints = NULL
)
}
\description{
Run BFGS Step with Analytic Gradients
}
\keyword{internal}
