% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_to_matrices}
\alias{theta_to_matrices}
\title{Convert theta vector back to R and S matrices}
\usage{
theta_to_matrices(theta, p)
}
\description{
Convert theta vector back to R and S matrices
}
\keyword{internal}
